refactor(server): Move Result to zally-core #560
