+++ server/src/main/java/de/zalando/zally/apireview/ApiReview.kt
@@ -1,7 +1,7 @@
 package de.zalando.zally.apireview
 
 import de.zalando.zally.dto.ApiDefinitionRequest
-import de.zalando.zally.rule.Result
+import de.zalando.zally.core.Result
 import de.zalando.zally.rule.api.Severity
 import org.hibernate.annotations.Parameter
 import org.hibernate.annotations.Type
