+++ server/src/main/java/de/zalando/zally/dto/ViolationsCounter.kt
@@ -1,6 +1,6 @@
 package de.zalando.zally.dto
 
-import de.zalando.zally.rule.Result
+import de.zalando.zally.core.Result
 import de.zalando.zally.rule.api.Severity
 
 class ViolationsCounter(violations: List<Result>) {
