+++ server/src/main/java/de/zalando/zally/rule/ApiValidator.kt
@@ -1,5 +1,6 @@
 package de.zalando.zally.rule
 
+import de.zalando.zally.core.Result
 import de.zalando.zally.core.RulesPolicy
 
 interface ApiValidator {
