+++ server/src/main/java/de/zalando/zally/rule/RulesValidator.kt
@@ -10,6 +10,7 @@ import de.zalando.zally.core.ContentParseResult.ParsedWithErrors
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
 import de.zalando.zally.core.JsonPointerLocator
+import de.zalando.zally.core.Result
 import de.zalando.zally.core.RulesManager
 import de.zalando.zally.core.RulesPolicy
 import org.slf4j.LoggerFactory
