+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/Result.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import com.fasterxml.jackson.core.JsonPointer
 import de.zalando.zally.rule.api.Severity
